Make index.js more readable
