+++ src/components/Header.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import Helmet from 'react-helmet';
+
+const Header = props => (
+    <Helmet>
+        <meta charSet="utf-8" />
+        <title>Tim Trautman</title>
+    </Helmet>
+);
+
+export default Header;
